+++ VoiceIt2-IosSDK/Classes/VideoEnrollmentViewController.m
@@ -232,6 +232,21 @@ float initialBrightnessVE = 0.0;
 }
 
 -(void)startEnrollmentProcess {
+    
+    //Check if User Id is correct
+    if (![Utilities checkUserId:self.userToEnrollUserId]){
+        NSString *response = @"{\"responseCode\":\"CREDENTIAL_ERROR\",\"message\":\"Please make sure your userId and credentials are correct\"}";
+        [self exitWithResponse:response];
+        return;
+    }
+    
+    //Check if there is Internet Connection
+    if (![Utilities checkNetwork]){
+        NSString *response = @"{\"responseCode\":\"NETWORK_ERROR\",\"message\":\"Please make sure you are connected to the internet\"}";
+        [self exitWithResponse:response];
+        return;
+    }
+    
     [self.myVoiceIt deleteAllEnrollments:self.userToEnrollUserId callback:^(NSString * deleteEnrollmentsJSONResponse){
         [self makeLabelFlyIn: [ResponseManager getMessage:@"GET_ENROLLED"]];
         [self startDelayedRecording:2.0];
@@ -310,6 +325,21 @@ float initialBrightnessVE = 0.0;
     });
 }
 
+-(void) exitWithResponse:(NSString *) jsonResponse {
+    NSLog(@"%@", jsonResponse);
+    NSDictionary *jsonObj = [Utilities getJSONObject:jsonResponse];
+    NSString * responseCode = [jsonObj objectForKey:@"responseCode"];
+    dispatch_async(dispatch_get_main_queue(), ^{
+        [self.messageLabel setText: [ResponseManager getMessage: responseCode]];
+        [self.messageLabel setAdjustsFontSizeToFitWidth:YES];
+    });
+    dispatch_after(dispatch_time(DISPATCH_TIME_NOW, 3.0 * NSEC_PER_SEC), dispatch_get_main_queue(), ^{
+        [[self navigationController] dismissViewControllerAnimated:YES completion:^{
+            [[self myNavController] userEnrollmentsCancelled];
+        }];
+    });
+}
+
 #pragma mark - Camera Delegate Methods
 
 // Code to Capture Face Rectangle and other cool metadata stuff
@@ -353,7 +383,18 @@ didOutputSampleBuffer:(CMSampleBufferRef)sampleBuffer
     if(self.imageNotSaved){
         // Convert to CIPixelBuffer for faceDetector
         CVPixelBufferRef pixelBuffer = CMSampleBufferGetImageBuffer(sampleBuffer);
-        if (pixelBuffer == NULL) { return; }
+        if (pixelBuffer == NULL) {
+            // Programmatically create white image to send
+            CGSize size = CGSizeMake(640, 480);
+            UIGraphicsBeginImageContextWithOptions(size, YES, 0);
+            [[UIColor whiteColor] setFill];
+            UIRectFill(CGRectMake(0, 0, size.width, size.height));
+            UIImage *uimage = UIGraphicsGetImageFromCurrentImageContext();
+            UIGraphicsEndImageContext();
+            CIImage* image = [[CIImage alloc] initWithCGImage:uimage.CGImage];
+            [self saveImageData:image];
+            return;
+        }
         
         // Create CIImage for faceDetector
         CIImage *image = [CIImage imageWithCVImageBuffer:pixelBuffer];
