+++ VoiceIt2-IosSDK/Classes/VoiceEnrollmentViewController.m
@@ -128,6 +128,21 @@ float initialBrightnessVoiceE = 0.0;
 }
 
 -(void)startEnrollmentProcess {
+    
+    //Check if User Id is correct
+    if (![Utilities checkUserId:self.userToEnrollUserId]){
+        NSString *response = @"{\"responseCode\":\"CREDENTIAL_ERROR\",\"message\":\"Please make sure your userId and credentials are correct\"}";
+        [self exitWithResponse:response];
+        return;
+    }
+    
+    //Check if there is Internet Connection
+    if (![Utilities checkNetwork]){
+        NSString *response = @"{\"responseCode\":\"NETWORK_ERROR\",\"message\":\"Please make sure you are connected to the internet\"}";
+        [self exitWithResponse:response];
+        return;
+    }
+    
     [self.myVoiceIt deleteAllEnrollments:self.userToEnrollUserId callback:^(NSString * deleteEnrollmentsJSONResponse){
                 NSLog(@"DELETING ENROLLMENTS IN THE BEGINNING %@",deleteEnrollmentsJSONResponse);
                 [self startDelayedRecording:0.0];
@@ -188,6 +203,21 @@ float initialBrightnessVoiceE = 0.0;
             [[self myNavController] userEnrollmentsCancelled];
         }];
         [self.myVoiceIt deleteAllEnrollments:self.userToEnrollUserId callback:^(NSString * deleteEnrollmentsJSONResponse){}];
+    });
+}
+
+-(void) exitWithResponse:(NSString *) jsonResponse {
+    NSLog(@"%@", jsonResponse);
+    NSDictionary *jsonObj = [Utilities getJSONObject:jsonResponse];
+    NSString * responseCode = [jsonObj objectForKey:@"responseCode"];
+    dispatch_async(dispatch_get_main_queue(), ^{
+        [self.messageLabel setText: [ResponseManager getMessage: responseCode]];
+        [self.messageLabel setAdjustsFontSizeToFitWidth:YES];
+    });
+    dispatch_after(dispatch_time(DISPATCH_TIME_NOW, 3.0 * NSEC_PER_SEC), dispatch_get_main_queue(), ^{
+        [[self navigationController] dismissViewControllerAnimated:YES completion:^{
+            [[self myNavController] userEnrollmentsCancelled];
+        }];
     });
 }
